requirements.txt
pandasai
bs4 # necessary for pandasai?
python-dotenv